+++ packages/devtools_shared/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 8.2.0
+* Use DTD in the `packageRootFromFileUriString` utility to detect the package root.
+
+# 8.1.1
+* TODO: publish this version with min Dart SDK version of the next Dart Beta.
+
 # 8.1.1-dev.0
 * Update `package:dtd` to `2.1.0`.
 * Bump the minimum Dart SDK version to `3.4.0-275.0.dev`.